Rename rule; add U case
