+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -43,7 +43,7 @@ pub(crate) use nonlocal_without_binding::*;
 pub(crate) use property_with_parameters::*;
 pub(crate) use redefined_argument_from_local::*;
 pub(crate) use redefined_loop_name::*;
-pub(crate) use redundant_u_string_prefix::*;
+pub(crate) use redundant_unicode_prefix::*;
 pub(crate) use repeated_equality_comparison::*;
 pub(crate) use repeated_isinstance_calls::*;
 pub(crate) use repeated_keyword_argument::*;
@@ -122,7 +122,7 @@ mod nonlocal_without_binding;
 mod property_with_parameters;
 mod redefined_argument_from_local;
 mod redefined_loop_name;
-mod redundant_u_string_prefix;
+mod redundant_unicode_prefix;
 mod repeated_equality_comparison;
 mod repeated_isinstance_calls;
 mod repeated_keyword_argument;
